# perf: reduce allocations of DoMultiCache