--- a/internal/util/parallel.go
+++ b/internal/util/parallel.go
@@ -20,25 +20,24 @@ func ParallelVals[K comparable, V any](maxp int, p map[K]V, fn func(k V)) {
 	closeThenParallel(maxp, ch, fn)
 }
 
+func worker[V any](wg *sync.WaitGroup, ch chan V, fn func(k V)) {
+	for v := range ch {
+		fn(v)
+	}
+	wg.Done()
+}
+
 func closeThenParallel[V any](maxp int, ch chan V, fn func(k V)) {
 	close(ch)
 	concurrency := len(ch)
 	if concurrency > maxp {
 		concurrency = maxp
 	}
-	wg := sync.WaitGroup{}
+	var wg sync.WaitGroup
 	wg.Add(concurrency)
 	for i := 1; i < concurrency; i++ {
-		go func(wg *sync.WaitGroup) {
-			for v := range ch {
-				fn(v)
-			}
-			wg.Done()
-		}(&wg)
+		go worker(&wg, ch, fn)
 	}
-	for v := range ch {
-		fn(v)
-	}
-	wg.Done()
+	worker(&wg, ch, fn)
 	wg.Wait()
 }