```
▶ benchstat old.txt new.txt
name                                 old time/op    new time/op    delta
/OneNode/128_RueidisDoMulti-10         42.0µs ± 0%    42.1µs ± 0%   +0.33%  (p=0.000 n=20+19)
/OneNode/128_RueidisDoMultiCache-10    4.24µs ± 1%    3.96µs ± 2%   -6.73%  (p=0.000 n=19+20)
/Cluster/128_RueidisDoMulti-10         28.8µs ± 4%    29.1µs ± 4%     ~     (p=0.157 n=20+20)
/Cluster/128_RueidisDoMultiCache-10    4.24µs ± 2%    4.21µs ± 3%   -0.86%  (p=0.037 n=20+20)

name                                 old alloc/op   new alloc/op   delta
/OneNode/128_RueidisDoMulti-10         18.5kB ± 0%    18.5kB ± 0%     ~     (p=0.513 n=20+19)
/OneNode/128_RueidisDoMultiCache-10    11.8kB ± 1%    11.5kB ± 0%   -2.84%  (p=0.000 n=20+20)
/Cluster/128_RueidisDoMulti-10         19.4kB ± 1%    19.0kB ± 1%   -1.80%  (p=0.000 n=20+20)
/Cluster/128_RueidisDoMultiCache-10    11.8kB ± 1%    11.2kB ± 1%   -4.76%  (p=0.000 n=20+20)

name                                 old allocs/op  new allocs/op  delta
/OneNode/128_RueidisDoMulti-10            130 ± 0%       130 ± 0%     ~     (all equal)
/OneNode/128_RueidisDoMultiCache-10      17.0 ± 0%      10.0 ± 0%  -41.18%  (p=0.000 n=18+20)
/Cluster/128_RueidisDoMulti-10            141 ± 0%       135 ± 0%   -4.26%  (p=0.000 n=20+20)
/Cluster/128_RueidisDoMultiCache-10      16.0 ± 0%      10.0 ± 0%  -37.50%  (p=0.000 n=20+20)
```

source
```go
package sep

import (
	"context"
	"github.com/redis/rueidis"
	"math/rand"
	"strconv"
	"testing"
	"time"
)

var charset = []byte("abcdefghijklmnopqrstuvwxyz")

func randstr(n int) string {
	b := make([]byte, n)
	for i := range b {
		b[i] = charset[rand.Intn(len(charset))]
	}
	return string(b)
}

func Benchmark(b *testing.B) {
	testfn := func(b *testing.B, n int, addresses []string) {
		ns := strconv.Itoa(n)
		makeclient := func(addresses []string) rueidis.Client {
			client, err := rueidis.NewClient(rueidis.ClientOption{
				InitAddress: addresses,
			})
			if err != nil {
				panic(err)
			}
			return client
		}

		// prepare keys
		keys := make(map[string]string, n)
		cmds := make(rueidis.Commands, 0, n)

		{
			client := makeclient(addresses)
			for i := 0; i < n; i++ {
				keys[randstr(10)] = randstr(50)
			}
			for k, v := range keys {
				cmds = append(cmds, client.B().Set().Key(k).Value(v).Build())
			}
			if err := client.Do(context.Background(), client.B().Flushall().Build()).Error(); err != nil {
				panic(err)
			}
			resps := client.DoMulti(context.Background(), cmds...)
			if len(resps) != len(keys) {
				panic("worn")
			}
			client.Close()
		}
		b.Run(ns+" RueidisDoMulti", func(b *testing.B) {
			client := makeclient(addresses)
			defer client.Close()
			commands := make([]rueidis.Completed, 0, len(keys))
			for k := range keys {
				commands = append(commands, client.B().Get().Key(k).Build().Pin())
			}
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				for pb.Next() {
					ret := client.DoMulti(context.Background(), commands...)
					for _, result := range ret {
						_, err := result.ToString()
						if err != nil {
							panic(err)
						}
					}
				}
			})
			b.StopTimer()
		})
		b.Run(ns+" RueidisDoMultiCache", func(b *testing.B) {
			client := makeclient(addresses)
			defer client.Close()
			commands := make([]rueidis.CacheableTTL, 0, len(keys))
			for k := range keys {
				commands = append(commands, rueidis.CT(client.B().Get().Key(k).Cache().Pin(), time.Second*3))
			}
			b.ResetTimer()
			b.ReportAllocs()
			b.RunParallel(func(pb *testing.PB) {

				for pb.Next() {
					ret := client.DoMultiCache(context.Background(), commands...)
					for _, result := range ret {
						_, err := result.ToString()
						if err != nil {
							panic(err)
						}
						//if v != keys[commands[i].Cmd.Commands()[1]] {
						//	panic("wrong")
						//}
					}
				}
			})
			b.StopTimer()
		})
	}

	b.Run("OneNode", func(b *testing.B) {
		for _, n := range []int{128} {
			testfn(b, n, []string{"127.0.0.1:6379"})
		}
	})

	b.Run("Cluster", func(b *testing.B) {
		for _, n := range []int{128} {
			testfn(b, n, []string{"127.0.0.1:7001", "127.0.0.1:7002", "127.0.0.1:7003"})
		}
	})
}
```